Deleted splitView.js and added the code to layoutMode.js
